Make `--ignore-python` the default for invocation as `aloe`, not `nosetests`.
